First working pass at direct connection mutations
Within a transaction the state of connections are expanded so that
fields represent the documents directly. After the transaction the
documents are collapsed into their references.

It might be that expansion/collapsing is not necessary as long as
documents are created and stored by their type with validations to
ensure consistency after a transaction. The connect, disconnect
operations in this commit has been modified to replicate what would
be done by the code in the transaction but ultimately these two
operations can be removed.

Additional tests will need to be added to check that expansion and
collapsing reflects what is allowed given the schema with objects,
lists, and non-nulls.

In the future if something should not be represented as a document
but live in on a connection field there could possibly be some
configuration provided to specify which fields should not be
treated as connections, but my guess is that this would be a niche
nice-to-have case.