Remove getDocumentsForType in favor of getStore
getStore allows direct access to the store and getting access to
the documents for a type is no longer accessible since:
getStore()[typename] offers the same access.

This commit also freezes access on pushing to the store on the
publicly provided frozen version of the store on the data property
of a Paper instance.